+++ Observer/CreditmemoSaveAfterObserver.php
@@ -104,7 +104,7 @@ class CreditmemoSaveAfterObserver implements ObserverInterface
             $response = $this->webService->createRefund($requestData);
             if (in_array($response['code'], ProntoPagaHelper::STATUS_OK)) {
                 if ($response['body']['status'] === ProntoPagaHelper::STATUS_SUCCESS) {
-                    // $this->updateTransaction($transaction);
+                    $this->updateTransaction($transaction);
                     $message = __('Refund confirmed by Pronto Paga, The credit memo was successfully refunded. Reference: %1', $creditmemo->getIncrementId());
                     $order->addCommentToStatusHistory($message);
                     $this->orderRepository->save($order);
@@ -164,7 +164,9 @@ class CreditmemoSaveAfterObserver implements ObserverInterface
      */
     private function updateTransaction($transaction)
     {
-        $transaction->setStatus(ProntoPagaHelper::STATUS_REFUNDED);
-        $this->transactionInterface->save($transaction);
+        if ($this->prontoPagaHelper->localValidation()) {
+            $transaction->setStatus(ProntoPagaHelper::STATUS_REFUNDED);
+            $this->transactionInterface->save($transaction);
+        }
     }
 }