+++ CHANGELOG.md
@@ -8,7 +8,7 @@ CHANGELOG
 * Updated JS files and apply improvements by eslint.
 * Update UI listing to show request body an response payload parsed by html entity.
 
-### 1.0.2 (2024-04-14)
+### 1.0.2 (2024-04-23)
 * Two observer are created for when the credit memo of the order is generated and the refund is carried out through ProntoPaga.
 * Added configuration for the observer on system.xml
 * Update validate action from Pronto Paga transaction UI grid.
@@ -16,3 +16,6 @@ CHANGELOG
 * Added new values on config.xml.
 * Added new status `refuned`.
 * Fix css.
+* Rename function `validateOnCheckout` to `localValidation` to avoid confusion.
+* Encrypted params ared added to  callbackUrl to validate the request and check the step.
+